[#204] 스타일 가이드 적용, AppModel 생성
