+++ MoMo/MoMo/Resources/Storyboards/Onboarding.storyboard
@@ -1024,6 +1024,7 @@
     </scenes>
     <designables>
         <designable name="hue-vy-02C"/>
+        <designable name="rIQ-3J-AWL"/>
     </designables>
     <resources>
         <image name="btnIcCalender16" width="16" height="16"/>
