+++ MoMo/MoMo/Sources/APIModels/DiaryWithId.swift
