+++ MoMo/MoMo/Sources/APIModels/Emotion.swift
