+++ MoMo/MoMo/Sources/AppModels/AppSentence.swift
@@ -1,0 +1,19 @@
+//
+//  MoodSentence.swift
+//  MoMo
+//
+//  Created by 초이 on 2021/01/27.
+//
+
+import UIKit
+
+// TODO: - 해석이오빠가 리팩토링하면서 네이밍 변경하거나 없애기
+// Book, Quote,
+
+struct AppSentence {
+    var id: Int?
+    var author: String
+    var bookTitle: String
+    var publisher: String
+    var sentence: String
+}
