+++ MoMo/MoMo/Sources/ViewControllers/MoodViewController.swift
@@ -7,115 +7,6 @@
 
 import UIKit
 
-enum Mood: Int {
-    case love = 1, happy, console, angry, sad, bored, memory, daily
-    
-    func toString() -> String {
-        switch self {
-        case .love:
-            return "사랑"
-        case .happy:
-            return "행복"
-        case .console:
-            return "위로"
-        case .angry:
-            return "화남"
-        case .sad:
-            return "슬픔"
-        case .bored:
-            return "우울"
-        case .memory:
-            return "추억"
-        case .daily:
-            return "일상"
-        }
-    }
-    
-    func toIcon() -> UIImage {
-        switch self {
-        case .love:
-            return Constants.Design.Image.icLove14Black!
-        case .happy:
-            return Constants.Design.Image.icHappy14Black!
-        case .console:
-            return Constants.Design.Image.icConsole14Black!
-        case .angry:
-            return Constants.Design.Image.icAngry14Black!
-        case .sad:
-            return Constants.Design.Image.icSad14Black!
-        case .bored:
-            return Constants.Design.Image.icBored14Black!
-        case .memory:
-            return Constants.Design.Image.icMemory14Black!
-        case .daily:
-            return Constants.Design.Image.icDaily14Black!
-        }
-    }
-    
-    func toWhiteIcon() -> UIImage {
-        switch self {
-        case .love:
-            return Constants.Design.Image.icLove14White!
-        case .happy:
-            return Constants.Design.Image.icHappy14White!
-        case .console:
-            return Constants.Design.Image.icConsole14White!
-        case .angry:
-            return Constants.Design.Image.icAngry14White!
-        case .sad:
-            return Constants.Design.Image.icSad14White!
-        case .bored:
-            return Constants.Design.Image.icBored14White!
-        case .memory:
-            return Constants.Design.Image.icMemory14White!
-        case .daily:
-            return Constants.Design.Image.icDaily14White!
-        }
-    }
-    
-    func toBlueIcon() -> UIImage {
-        switch self {
-        case .love:
-            return Constants.Design.Image.icLove14Blue!
-        case .happy:
-            return Constants.Design.Image.icHappy14Blue!
-        case .console:
-            return Constants.Design.Image.icConsole14Blue!
-        case .angry:
-            return Constants.Design.Image.icAngry14Blue!
-        case .sad:
-            return Constants.Design.Image.icSad14Blue!
-        case .bored:
-            return Constants.Design.Image.icBored14Blue!
-        case .memory:
-            return Constants.Design.Image.icMemory14Blue!
-        case .daily:
-            return Constants.Design.Image.icDaily14Blue!
-        }
-    }
-    
-    func toTypingLabelText() -> String {
-        switch self {
-        case .love:
-            return "새로운 인연이 기대되는 하루였다. "
-        case .happy:
-            return "삶의 소중함을 느낀 하루였다. "
-        case .console:
-            return "나를 위한 진한 위로가 필요한 하루였다. "
-        case .angry:
-            return "끓어오르는 속을 진정시켜야 하는 하루였다. "
-        case .sad:
-            return "마음이 찌릿하게 아픈 하루였다. "
-        case .bored:
-            return "눈물이 왈칵 쏟아질 것 같은 하루였다. "
-        case .memory:
-            return "오래된 기억이 되살아나는 하루였다. "
-        case .daily:
-            return "평안한 하루가 감사한 날이었다."
-        }
-    }
-}
-
 struct Button {
     
     let button: UIButton?
@@ -381,7 +272,7 @@ class MoodViewController: UIViewController {
         }
     }
 
-    
+    // TODO: - AppDate로 빼기
     func weekdayEnglishToKorean(weekday: String) -> String {
         switch weekday {
         case "Monday":
@@ -425,42 +316,42 @@ class MoodViewController: UIViewController {
     }
     
     @IBAction func loveButtonTouchUp(_ sender: UIButton) {
-        pushToOnboardingSentenceViewController(mood: Mood.love, usage: changeUsage)
+        pushToOnboardingSentenceViewController(mood: AppEmotion.love, usage: changeUsage)
     }
     
     @IBAction func happyButtonTouchUp(_ sender: UIButton) {
-        pushToOnboardingSentenceViewController(mood: Mood.happy, usage: changeUsage)
+        pushToOnboardingSentenceViewController(mood: AppEmotion.happy, usage: changeUsage)
     }
     
     @IBAction func consoleButtonTouchUp(_ sender: UIButton) {
-        pushToOnboardingSentenceViewController(mood: Mood.console, usage: changeUsage)
+        pushToOnboardingSentenceViewController(mood: AppEmotion.console, usage: changeUsage)
     }
     
     @IBAction func angryButtonTouchUp(_ sender: UIButton) {
-        pushToOnboardingSentenceViewController(mood: Mood.angry, usage: changeUsage)
+        pushToOnboardingSentenceViewController(mood: AppEmotion.angry, usage: changeUsage)
     }
     
     @IBAction func sadButtonTouchUp(_ sender: UIButton) {
-        pushToOnboardingSentenceViewController(mood: Mood.sad, usage: changeUsage)
+        pushToOnboardingSentenceViewController(mood: AppEmotion.sad, usage: changeUsage)
     }
     
     @IBAction func boredButtonTouchUp(_ sender: UIButton) {
-        pushToOnboardingSentenceViewController(mood: Mood.bored, usage: changeUsage)
+        pushToOnboardingSentenceViewController(mood: AppEmotion.bored, usage: changeUsage)
     }
     
     @IBAction func memoryButtonTouchUp(_ sender: UIButton) {
-        pushToOnboardingSentenceViewController(mood: Mood.memory, usage: changeUsage)
+        pushToOnboardingSentenceViewController(mood: AppEmotion.memory, usage: changeUsage)
     }
     
     @IBAction func dailyButtonTouchUp(_ sender: UIButton) {
-        pushToOnboardingSentenceViewController(mood: Mood.daily, usage: changeUsage)
+        pushToOnboardingSentenceViewController(mood: AppEmotion.daily, usage: changeUsage)
     }
     
     @objc func touchCloseButton() {
         self.navigationController?.popViewController(animated: true)
     }
     
-    func pushToOnboardingSentenceViewController(mood: Mood, usage: Bool) {
+    func pushToOnboardingSentenceViewController(mood: AppEmotion, usage: Bool) {
         guard let date = dateLabel.text else {
             return
         }
